move better climate year details class to own file
