+++ src/app/account/account-reports/better-climate-report/better-climate-report.component.html
@@ -1,12 +1,12 @@
 <ng-container *ngIf="selectedReport && calculating == false">
     <div class="report" [ngClass]="{'print': print}">
-        <app-portfolio-information-table [yearDetails]="betterClimateReport.yearDetails" [cellWidth]="cellWidth"></app-portfolio-information-table>
-        <app-absolute-emissions-table [yearDetails]="betterClimateReport.yearDetails" [cellWidth]="cellWidth"></app-absolute-emissions-table>
-        <app-emissions-reductions-table [yearDetails]="betterClimateReport.yearDetails" [cellWidth]="cellWidth"></app-emissions-reductions-table>
-        <app-total-portfolio-energy-use-table [yearDetails]="betterClimateReport.yearDetails" [cellWidth]="cellWidth"></app-total-portfolio-energy-use-table>
+        <app-portfolio-information-table [yearDetails]="betterClimateReport.portfolioYearDetails" [cellWidth]="cellWidth"></app-portfolio-information-table>
+        <app-absolute-emissions-table [yearDetails]="betterClimateReport.portfolioYearDetails" [cellWidth]="cellWidth"></app-absolute-emissions-table>
+        <app-emissions-reductions-table [yearDetails]="betterClimateReport.portfolioYearDetails" [cellWidth]="cellWidth"></app-emissions-reductions-table>
+        <app-total-portfolio-energy-use-table [yearDetails]="betterClimateReport.portfolioYearDetails" [cellWidth]="cellWidth"></app-total-portfolio-energy-use-table>
         <!--TODO: Add vehicle emissions to meters-->
-        <!-- <app-vehicle-energy-use-table [yearDetails]="betterClimateReport.yearDetails" [cellWidth]="cellWidth"></app-vehicle-energy-use-table> -->
-        <app-graph-calculations-table [yearDetails]="betterClimateReport.yearDetails" [cellWidth]="cellWidth"></app-graph-calculations-table>
+        <!-- <app-vehicle-energy-use-table [yearDetails]="betterClimateReport.portfolioYearDetails" [cellWidth]="cellWidth"></app-vehicle-energy-use-table> -->
+        <app-graph-calculations-table [yearDetails]="betterClimateReport.portfolioYearDetails" [cellWidth]="cellWidth"></app-graph-calculations-table>
     </div>
 </ng-container>
