+++ src/app/account/account-reports/better-climate-report/emissions-reductions-table/emissions-reductions-table.component.ts
@@ -1,5 +1,5 @@
 import { Component, Input } from '@angular/core';
-import { BetterClimateYearDetails } from 'src/app/calculations/carbon-calculations/betterClimateReport';
+import { BetterClimateYearDetails } from 'src/app/calculations/carbon-calculations/betterClimateYearsDetails';
 
 @Component({
   selector: 'app-emissions-reductions-table',
