+++ src/app/account/account-reports/better-climate-report/graph-calculations-table/graph-calculations-table.component.ts
@@ -1,5 +1,5 @@
 import { Component, Input } from '@angular/core';
-import { BetterClimateYearDetails } from 'src/app/calculations/carbon-calculations/betterClimateReport';
+import { BetterClimateYearDetails } from 'src/app/calculations/carbon-calculations/betterClimateYearsDetails';
 
 @Component({
   selector: 'app-graph-calculations-table',
