+++ src/app/account/account-reports/better-climate-report/total-portfolio-energy-use-table/total-portfolio-energy-use-table.component.ts
@@ -1,5 +1,5 @@
 import { Component, Input } from '@angular/core';
-import { BetterClimateYearDetails } from 'src/app/calculations/carbon-calculations/betterClimateReport';
+import { BetterClimateYearDetails } from 'src/app/calculations/carbon-calculations/betterClimateYearsDetails';
 
 @Component({
   selector: 'app-total-portfolio-energy-use-table',
