+++ src/app/account/account-reports/better-climate-report/vehicle-energy-use-table/vehicle-energy-use-table.component.ts
@@ -1,5 +1,5 @@
 import { Component, Input } from '@angular/core';
-import { BetterClimateYearDetails } from 'src/app/calculations/carbon-calculations/betterClimateReport';
+import { BetterClimateYearDetails } from 'src/app/calculations/carbon-calculations/betterClimateYearsDetails';
 
 @Component({
   selector: 'app-vehicle-energy-use-table',
